requirements.txt
flask==3.0.3
pywebview==5.1
pyodbc==5.1.0
python-dotenv==1.0.1
werkzeug==3.0.3
jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
pyinstaller==6.8.0
